mariano20/requirements.txt
apipkg==1.4
beautifulsoup4==4.4.1
blessings==1.6
decorator==4.0.10
execnet==1.4.1
filemagic==1.6
flake8==2.6.2
ipdb==0.10.1
ipython==5.0.0
ipython-genutils==0.1.0
mccabe==0.5.0
pexpect==4.2.0
pickleshare==0.7.2
pkg-resources==0.0.0
prompt-toolkit==1.0.3
ptyprocess==0.5.1
py==1.4.31
pycodestyle==2.0.0
pyflakes==1.2.3
Pygments==2.1.3
pytest==2.9.1
pytest-colordots==0.1
pytest-flake8==0.5
pytest-sugar==0.7.1
pytest-xdist==1.14
requests==2.10.0
simplegeneric==0.8.1
six==1.10.0
termcolor==1.1.0
traitlets==4.2.2
wcwidth==0.1.7
apipkg==1.4
beautifulsoup4==4.4.1
blessings==1.6
decorator==4.0.10
execnet==1.4.1
filemagic==1.6
flake8==2.6.2
ipdb==0.10.1
ipython==5.0.0
ipython-genutils==0.1.0
mccabe==0.5.0
pexpect==4.2.0
pickleshare==0.7.2
pkg-resources==0.0.0
prompt-toolkit==1.0.3
ptyprocess==0.5.1
py==1.4.31
pycodestyle==2.0.0
pyflakes==1.2.3
Pygments==2.1.3
pytest==2.9.1
pytest-colordots==0.1
pytest-flake8==0.5
pytest-sugar==0.7.1
pytest-xdist==1.14
PyYAML==3.11
requests==2.10.0
simplegeneric==0.8.1
six==1.10.0
termcolor==1.1.0
traitlets==4.2.2
wcwidth==0.1.7
